requirements.txt
Django==4.2.7
python-dotenv==1.0.0
Pillow==10.1.0
django-ckeditor==6.7.0
django-cleanup==8.0.0
django-tinymce==3.6.0
whitenoise==6.6.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
django-debug-toolbar==4.2.0